refactor: Rename justd components folder.
