+++ .changeset/long-lobsters-invite.md
@@ -1,0 +1,5 @@
+---
+"@haus23/tipprunde-hinterhof": patch
+---
+
+Rename justd components folder.
