+++ apps/hinterhof/app/routes/dashboard.tsx
@@ -1,4 +1,4 @@
-import { Heading } from '#/components/ui';
+import { Heading } from '#/components/ui-justd';
 
 export default function DashboardRoute() {
   return <Heading>Dashboard</Heading>;
