+++ apps/hinterhof/app/routes/master-data/championships.tsx
@@ -1,4 +1,4 @@
-import { Heading } from '#/components/ui';
+import { Heading } from '#/components/ui-justd';
 
 export default function ChampionshipsRoute() {
   return <Heading>Turniere</Heading>;
